pyproject.toml
[tool.poetry]
name = "adk-tutorial"
version = "0.1.0"
description = "Google ADK Tutorial Project"
authors = ["User"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
google-adk = "^1.4.2"
jupyter = "^1.0.0"
python-dotenv = "^1.0.1"
litellm = "^1.72.9"
google-generativeai = "^0.8.5"
certifi = "^2024.7.4"

[tool.poetry.scripts]
format-json = "app.format:format_json"
format-all = "app.format:format_all"
ruff-check = "app.format:ruff_check"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
extend-select = ["I", "UP"]

[tool.ruff.format]
line-ending = "auto"
